Fix broken billing tests caused by year change (#153)
https://eaflood.atlassian.net/browse/WATER-4974

We've gone to run our acceptance tests and found that some have failed. All are connected to the switch into the next financial year, and these changes are for failing supplementary tests.

As we move away from 2022, the number of bills expected in supplementary bills runs will increase, whilst PRE-SROC will reduce. We're handling this with calculated assertions in some places, but we missed some.

This change fixes those omissions.