Convert format helper into Cypress commands
> Many first time users look at Cypress code and think it runs synchronously.

Its the reason why some folks vehemently take against Cypress because you don't write it like normal JavaScript.

But as we are using it, we're required to follow the 'cypress' way. This means we shouldn't have modules full of vanilla functions. Instead, we need to use [commands](https://docs.cypress.io/app/core-concepts/introduction-to-cypress#Chains-of-Commands) that [wrap](https://docs.cypress.io/api/commands/wrap) their return values, so we can be sure they are run at the right time, instead of when they are first invoked.