Fix up historic 2 with all the changes made
Historic 2 was suffering from the same issues as found in historic-1 and historic-3, but had the added complexity of return logs covering multiple years and cycles.

We also found an issue with how it was seeding the form logs.

Assuming you had fixed the whole JavaScript month index thing, and April wasn't being considered as needing a year deducted, there was still an issue.

The logic for decrementing the year only kicked in when the current date was Jan, Feb or March for winter, and Jan to Oct for summer.

The following shows the ID's that were generated for the four return logs, when the date was 2025-04-03.

- 0 - WINTER - v1:9:AT/CURR/DAILY/01:9999990:2024-04-01:2025-03-31
- 1 - SUMMER - v1:9:AT/CURR/DAILY/01:9999991:2024-11-01:2025-10-31
- 2 - WINTER - v1:9:AT/CURR/DAILY/01:9999990:2023-04-01:2024-03-31
- 3 - SUMMER - v1:9:AT/CURR/DAILY/01:9999991:2023-11-01:2024-10-31

You can see they are all unique. Contrast that with when the date was 2024-08-27

- 0 - WINTER - v1:9:AT/CURR/DAILY/01:9999990:2024-04-01:2025-03-31
- 1 - SUMMER - v1:9:AT/CURR/DAILY/01:9999991:2023-11-01:2024-10-31
- 2 - WINTER - v1:9:AT/CURR/DAILY/01:9999990:2024-04-01:2025-03-31
- 3 - SUMMER - v1:9:AT/CURR/DAILY/01:9999991:2022-11-01:2023-10-31

You can see the return  ID's generated for the winter return logs are the same, which caused the seeding to error.