Remove now unused lib functions
We have been setting start, end, due and cycle dates all based on the current financial year, and then adjusting for it in our tests.

Unlike water-abstraction-system itself, as we control the data being seeded we shouldn't need to copy all the logic required to determine what values we want to use when seeding or verifying the data in our acceptance tests.