Update Return version setup note messages
https://eaflood.atlassian.net/browse/WATER-4869

We're replacing the legacy return log submission process. As part of that we are adding the ability to add a note modelled on how we do it for return versions.

But our UCD team want specific messages for the return log rather than the generic 'Changes made' we use in return versions.

That's cool, but it means we need to update the existing return versions to keep them aligned.

We did that in [Update return versions check page banners](https://github.com/DEFRA/water-abstraction-system/pull/1879) so this updates the return version test to match.