Update SROC based 2PT bill run tests for 2024/25 (#164)
https://eaflood.atlassian.net/browse/WATER-5064

In [Update Two-part tariff annual years for 2024-25](https://github.com/DEFRA/water-abstraction-system/pull/2002), we updated the years shown in the bill run setup journey's "Select the financial year" page to reflect that the SROC 2PT bill runs have been run for 2022/23 and 2023/24. On the backlog are just 2020/21 and 2021/22. For future years, it will just be the previous financial year (currently 2024/25).

The same changes need to be made to the SROC-based two-part tariff bill run acceptance tests, as they now all fail.