Update return log tests after reference change
https://eaflood.atlassian.net/browse/WATER-5085

An issue was found with the way we were generating the 'legacy_id' (return reference) for new return requirements. We had incorrectly assumed that they were like most NALD records: unique within a region only.

So, our logic would determine the current largest reference within a region, add 1 to it, and use that for the new reference.

Turns out since 2008 they have been unique across _all_ regions. So now we have an auto-increment on the table itself. When new records are added PostgreSQL just sets the reference to the next increment.

The issue for the tests is, we are assigning a legacy ID that is in the 999999* range to avoid clashing with real return requirements. But now the reference is determined by the DB, it will just look at what its next increment will be. At time of writing, for all environments that is in the 10****** range.

When it comes to the historic return log tests, they are still seeing the expected return logs created, its just they are now in a different order because of the change in references.