Update return log tests after reference change (#165)
https://eaflood.atlassian.net/browse/WATER-5085

An issue was identified with the method used to generate the 'legacy_id' (return reference) for new return requirements. We had incorrectly assumed that they were like most NALD records: unique within a region only.

Our logic would determine the maximum reference within a region, add 1 to it, and use that for the new reference.

It turns out that, since 2008, they have been unique across _all_ regions. So, we have updated the service to utilise a SQL auto-increment built into the table itself. When new records are added, PostgreSQL sets the reference to the next increment.

The issue with the tests is that we are assigning a legacy ID in the 999999* range to avoid clashing with real return requirements. However, now the reference is determined by the DB, so it will simply look at what its next increment will be. At the time of writing, for all environments that is in the 10****** range.

When it comes to the historic return log tests, they are still seeing the expected return logs created; it's just that they are now in a different order due to the change in references.