Fix test broken by changes to return version setup
https://eaflood.atlassian.net/browse/WATER-5096

We got return version setup wrong by basing all the abstraction data we show in the journey on the 'current' licence version, even if the start date for the new return version is before it starts.

Now the journey selects the relevant licence version based on the start date entered. Another change is it having determined the relevant licence version, it will only allow users to copy from return versions whose start date falls in the relevant licence version's period.

It's this last change which has broken our historic correction tests. They are assuming the `Copy from existing` option will be available, but because of the fixtures that is not the case.

A minor tweak, and all is working again.