Update returns tests to match change to badges (#168)
https://eaflood.atlassian.net/browse/WATER-5006

We made a change to **water-abstraction-ui** to [Use new link and tag for returns in search results](https://github.com/DEFRA/water-abstraction-ui/pull/2714).

This meant we made some changes to the badges that the legacy UI displays.

One of the changes was to stop wasting time Title casing the text for the badges as the GOV.UK elements display the text in all uppercase anyway. This broke `cypress/e2e/internal/returns/record-receipt.cy.js`, hence a minor tweak is needed.

The other was to display 'not due yet' instead of 'due' for return logs that are more than 28 days before their due date. This broke `cypress/e2e/external/view-returns.cy.js`.