Update acceptance tests after refactor of views (#171)
https://eaflood.atlassian.net/browse/WATER-5139

Now that we have completed the return version set up journey it was time to revisit the views and standardise the code across them. As multiple people were working on the different pages it is always the case that there will be minor variations in how things are implemented.

This PR updates the acceptance tests based on the new updated views.