Update billing test due to changed charges (#181)
We know there are uplifts in charges planned for this year, and we have a test that asserts a value is equal to `x`.

Really, it should just be testing a value exists rather than the specific charge because these do change each year. So, this change updates the test to avoid it failing again in future years.