Fix billing tag selectors post 'template' std
https://eaflood.atlassian.net/browse/WATER-5202

We have recently been working through some of our views updating them to a 'template' standard. This is to ensure consistency in the service as [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) takes over more and more of the service.

One of the changes is dropping unnecessary applications of the `.govuk-body` class to elements. This is declared at the body level and only in a few exceptional cases does it need to be applied to an element. However, this means a number of selectors in the billing pages for the status tag need updating.