Update tests for monitoring stations and alerts (#184)
https://eaflood.atlassian.net/browse/WATER-5164

Break the existing test into 2. Add a new internal/monitoring-stations folder. Inside it create two folders.

- abstraction-alerts
- licence-tagging

The current test relies on the ‘tag a licence’ step in order to then test the abstraction alert and remove tag journeys. So, we also need to create one or more fixtures for these so we can skip straight to creating the alert and removing the tag https://eaflood.atlassian.net/browse/WATER-5163.