Fix cancel and send alert tests https://eaflood.atlassian.net/browse/WATER-5164
During QA an issue has been found with the "Send" and "Cancel" alert tests.

On the "Check the recipients" page the tests where checking for the "Recipient" email address "external@example.co.uk", but the tests were returning the email address "external@example.com".

Upon investigation I found that there was some erroneous data in my `licence_entities` view which was preventing the test data from seeding properly. Once removed the data was seeded correctly causing my tests to also fail locally on the email address in the same way they did during QA.

So in this PR the test for the recipient email is going to be corrected to check for the right email address.