Remove Notify callback test
https://eaflood.atlassian.net/browse/WATER-4976

Back in June 2025 we took the decision to remove the config in [GOV.UK Notify](https://www.notifications.service.gov.uk/) which enabled callbacks.

When **Notify** confirms delivery of a Notification it can ping a designated endpoint with the details. In theory this is better than us manually polling **Notify** to check the status of pending notifications. We get a 'real time' update and we only process the ones we know have been confirmed.

In practice **Notify** doesn't distinguish between live and test notifications, and we can only provide one URL. This means _any_ notification sent to Notify from any of our environments would result in Notify pinging our callback endpoint in `production`.

It in turn would then attempt to look up the notification and 9 times out of 10 find no match, with all the associated error logging that goes with that. There have been times when testing that we have inadvertently degraded the live service because we've caused **Notify** to essentially DOS the service!

Hence, we removed the config and stick with polling Notify on a nightly basis. But all this means that the **Notify** callback endpoint is now defunct which means so is the test.