Fix broken tests (#191)
The changes made in https://eaflood.atlassian.net/browse/WATER-5213 have changed the ordering on the Licence's return tab. As a result there are some tests which are now failing.

In this PR the "View returns and their status" and the "Editing a return" tests will be updated to work with the new ordering.