Update abstraction alerts show all recipients step (#193)
* Update abstraction alerts show all recipients step
We recently updated our layout to provided another block 'pageContent'. When this was done we did a little clean up of stray (and unecessary) '.govuk-body' divs.

This change updates the test selector to check there is a <p> on the page containing ''Showing all 1 recipients'.

The tests are the ame as before just an updated selector