Update broken tests after recent changes
https://eaflood.atlassian.net/browse/WATER-5212
https://eaflood.atlassian.net/browse/WATER-5281

We've made two changes to the service recently that the acceptance tests need to be updated to take account of.

The first is that we will shortly not be applying a due date to newly created return logs. We have a test that is still asserting a due date is being set.

We have also updated our send notice process, to automatically check the status of email notifications once we've sent them to Notify. Again, this means an assertion is failing because it is expecting `PENDING` when it could be `SENT` (or even `ERROR` if Notify is down).

This change updates those broke tests.