Update broken tests after recent changes (#197)
https://eaflood.atlassian.net/browse/WATER-5212
https://eaflood.atlassian.net/browse/WATER-5281

We've made two recent changes to the service that require the acceptance tests to be updated to account for them.

The first is that we will soon no longer apply a due date to newly created return logs. We have a test that is still asserting that a due date is being set.

We have also updated our send notice process to automatically check the status of email notifications once they have been sent to Notify. Again, this means an assertion is failing because it is expecting `PENDING` when it could be `SENT` (or even `ERROR` if Notify is down).

This change updates those broken tests.