Update licence alias naming test
This had some embedded tests we weren't aware of. These act as separate tests and cause cypress to start again. We can carry out the same checks as part of a single test, and thus avoid the overhead and complexity of asking Cypress to run nested tests.

The test itself just needs a login and a licence, so we reuse the sharing scenario data as it also works for this test.