Update the return statuses tests
We've moved it over to the new data loading mechanism and programmatic login. But we've also created it its own scenario data.

The previous test was simply confirming some of the statuses we expect were on the page somewhere. Now we load return logs with the intent of generating specific statuses on the page that we can then check for.