Update internal reset password test
To start with, the same argument regards validation applies here as we made for the external reset password validation test.

With that gone, rather than keeping it in its own folder it makes more sense to move it with the rest of the user/auth based tests.

In the test itself, we don't need to seed an additional user to run this scenario. This means we can drop the call to tear down, and just use the basic user all environments are seeded with.