Update paper returns tests
So we know this are not going to be around for long, as we are due to ship a replacement version of this journey. However, we'll still want test coverage afterwards.

So, we keep the journey as an indicator of that. But we drop validations, namely because it requires us to load a whole other licence in order to meet the assertions specified.

In the process of updating this test to drop barebones and use programmatic login, we found our test data was causing an error for the legacy code.

It relied on the region ID embedded in the metadata along with the format ID to match to a return requirement. Because we were setting the return requirement to use region 9 but the return log to use region 1, no match was being found so the legacy journey was erroring.