Add new calculatedDates command
To allow us to properly populate the service for the returns invite and reminder tests, we need to know what the current return periods are.

That's no simple matter, and would be impractical to duplicate the code here. So, instead, we've built a new endpoint in water-abstraction-system to provide those values to us.

Our new calculatedDates() command can be used to fetch them from the service so they can then be used in tests.