Make return statuses reusable
The external test needs a user as well. But the rest of what return-statuses was doing we really need in order to update the returns view-status test.

So, we move out most of the work into return-statuses.js which we can use directly in the view-status test, and use indirectly in the external return-statuses test.