Update all programmatic logins to be direct
We know the idea of using programmatic logins is to allow you to go straight to the page/function being tested.

We just wanted to make sure things were running and working on our first pass. Now we know they are we felt brave enough to update all the tests we'd switched to use programmatic login to also go direct.