Switch from AT/CURR/DAILY/01 to AT/TEST/01
AT/CURR/DAILY/01 was the go to it seems in the legacy tests we inherited, but it was called daily because it was linked to being setup for daily returns. This made sense when you were always seeding a daily licence, a weekly licence and a monthly licence.

But now where possible we just seed a single licence, and then whatever connected data we need. So, not to confuse issues this replaces its use with AT/TEST/01, which is the pattern used in the billing tests.