Update to use new View notice page (#201)
https://eaflood.atlassian.net/browse/WATER-4992

As part of migrating the legacy pages to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system), we have previously replaced the view notices page. Now it's the turn of the view notice page, which is linked to when a user clicks a notice to view.

This change updates any tests that are affected by it.