Add changing return cycle type screanrio to return logs (#200)
https://eaflood.atlassian.net/browse/WATER-5302

A user found an issue when making a historic change to a return version. When swapping return cycle type (from summer to all year) it was noticed that we were not voiding the existing return logs for the previous type.

This PR adds a test scenario for that example now that we have fixed that issue.