Fix return log tests broken by view updates (#203)
https://eaflood.atlassian.net/browse/WATER-5153

In PR https://github.com/DEFRA/water-abstraction-system/pull/2213 the views have been updated for the `return-logs/setup` journey to get them up to our "template" standard.

Part of this update involved changing some of the ids in the views. As the acceptance tests rely on these ids several of them now fail. These failing tests will be fixed in this PR.