Return log tests summer cycle (#205)
We recently swapped into a new summer return cycle (1st of November) which has highlighted that some of our return log tests that use both summer and all year cycles don't account for this change.

We need to expect different results on the return page during November and March. When the new all year cycle starts in April the tests will then change again. This PR adds in an if clause to handle this change over.