Return Log submission tests https://eaflood.atlassian.net/browse/WATER-5388
We initially didn’t spot that there were issues submitting return logs with no values, or return logs with blank lines.

We fixed them, only for the a regression issue to happen we didn’t spot. If we had acceptance test coverage, we would have spotted these issues.

So in this PR new tests are going to be created to cover all the ways a return log can be submitted or edited to give us full coverage of the return-log setup journey.