Fix selector for recalculate bills test (#209)
https://eaflood.atlassian.net/browse/WATER-5372

In [Update the licence mark and marked supplementary pages to the template standard](https://github.com/DEFRA/water-abstraction-system/pull/2637), we updated the pages used in the recalculate bills journey to "template standard".

However, this resulted in a change in the attributes of a link on the confirmation page that the test selects.

This updates the selector for it.