Update licence based tests to secondary navigation (#211)
* Update any pages using the updated layout to infer the 'pageHeadingHtml' (#210)
https://eaflood.atlassian.net/browse/WATER-5372

We have been updating views to 'template standard' this has involved using a shared page heading on all the updated pages to have some consistency.

However, we have not been consistent in its implementation.

This change updates the failing tests from the recent changes to template standard

Pages no longer need to set the 'pageHeadingHtml' on the page, this will be set in the layout, this ensures consistency.

This has the effect of making every page have a 'h1' to assert rather then the label.

https://github.com/DEFRA/water-abstraction-system/pull/2643