Enable use of new search page
https://eaflood.atlassian.net/browse/WATER-5292

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system` app.

To avoid multiple redirects, any links that point to the old  `/licences` path have been updated to point to the correct root path `/` instead.
This change updates the aceptance tests to reflect that.

The licence numbers that were used in the tests do not meet the search rules, so this change updates them to fit those rules, allowing the tests to pass.

Also, the search page content is different, so this change updates references to button names and the content of the search result table.