Fix issue with period
The paper return and returns invitation ad-hoc journeys rely on the licence having a 'due' return log.

We think these were working before because the engine wasn't discounting ones where the end date is greater than or equal to the current date.

Now it ease, it looks like the test data being seeded is using the current return period, which means the end date will always be in the future.

So, we add new command that will take the current period and push it back by one. This will always result in the return log that is seeded having an end date prior to the current date i.e. it will be 'due' and the tests will pass again.