Update send notice journeys (#214)
https://eaflood.atlassian.net/browse/WATER-4991

Because of changes made to the service in support of dynamic due dates and how they now impact sending notices, we need to make some changes that require us to update the acceptance tests.

One of them was ensuring the ad-hoc journeys only consider 'due' return logs that can be submitted. So, those with an end date on or after the current date are not yet due.

It looks like the ad-hoc paper return and returns invitation tests were seeding return logs where this was not the case, because they were using the start and end dates for the current period.

So, we add a new command that takes the current period and pushes it back. This will always result in the return log that is seeded having an end date before the current date, i.e. it will be 'due', and the tests will pass again.