Fix tests broken by adding `return_log_id` (#216)
https://eaflood.atlassian.net/browse/WATER-5425

Following the addition of the `return_log_id`column to the `returns.versions` table (AKA `return_submissions` view) in PRs https://github.com/DEFRA/water-abstraction-returns/pull/462 & https://github.com/DEFRA/water-abstraction-system/pull/2791 some of the tests now fail due to the new ID not being in the fixtures.

This PR will fix the tests that the addition of the new column has broken.